+++ src/config/index.ts
@@ -1,0 +1,25 @@
+import HttpConfig from '@/config/http';
+
+export interface Config {
+  /** Request and http configuration */
+  http: HttpConfig;
+}
+
+export const defaultHttpConfig: HttpConfig = {
+  headers: {
+    'content-type': 'application/json;charset=UTF-8',
+  },
+  relink: true,
+  retry: 2,
+  timeout: 5000,
+  notify: false,
+  notifyError: false,
+  baseURL: 'https://api.lipoic.org',
+  messageDuration: 4000,
+};
+
+export const globalConfig: Config = {
+  http: defaultHttpConfig,
+};
+
+export default globalConfig;
